businessPartners filters
Previously, all users were stored in MongoDB. Now that we are allowing business partners to login using the Azure backend, all features that were tied to users had to work with this new backend.

So now things like filters, campaigns and offers will store a business partner id, and will check the logged in user for this id.